updated magistrate classes
